+++ packages/jest-jasmine2/src/PCancelable.js
