+++ packages/jest-jasmine2/src/expectationResultFactory.js
