+++ packages/jest-jasmine2/src/index.js
@@ -16,10 +16,10 @@ import type Runtime from 'jest-runtime';
 
 import path from 'path';
 import installEach from './each';
-import {installErrorOnPrivate} from './error_on_private';
+import {installErrorOnPrivate} from './errorOnPrivate';
 import {getCallsite} from 'jest-util';
 import JasmineReporter from './reporter';
-import {install as jasmineAsyncInstall} from './jasmine_async';
+import jasmineAsyncInstall from './jasmineAsyncInstall';
 
 const JASMINE = require.resolve('./jasmine/jasmine_light.js');
 
@@ -117,7 +117,7 @@ async function jasmine2(
   env.addReporter(reporter);
 
   runtime
-    .requireInternalModule(path.resolve(__dirname, './jest_expect.js'))
+    .requireInternalModule(path.resolve(__dirname, './jestExpect.js'))
     .default({
       expand: globalConfig.expand,
     });
