+++ packages/jest-jasmine2/src/isError.js
