+++ packages/jest-jasmine2/src/jasmine/Spec.js
@@ -33,11 +33,11 @@ WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 import {AssertionError} from 'assert';
 
-import ExpectationFailed from '../expectation_failed';
+import ExpectationFailed from '../ExpectationFailed';
 
-import expectationResultFactory from '../expectation_result_factory';
+import expectationResultFactory from '../expectationResultFactory';
 
-import assertionErrorMessage from '../assert_support';
+import assertionErrorMessage from '../assertionErrorMessage';
 
 export default function Spec(attrs: Object) {
   this.resultCallback = attrs.resultCallback || function() {};
