+++ packages/jest-jasmine2/src/pTimeout.js
