+++ packages/jest-jasmine2/src/treeProcessor.js
